refactor: alert message 재사용할 수 있도록 하기
